Address follow-up review comments on browser SSO
- Merge participant and session-context stores into store.go (single store impl)
- Unify SQL query-constant IDs under one SSO-SESS-NN prefix
- Build the graphbuilder SSO test flow inline; drop testdata/sso_flow.json
- Reuse the flow version captured at context load instead of re-fetching the
  flow in resolveActiveFlowVersion

Refs #3779
