--- backend/internal/flow/graphbuilder/testdata/sso_flow.json
@@ -1,127 +1,0 @@
-{
-    "name": "Default SSO Authentication Flow",
-    "handle": "default-sso-flow",
-    "flowType": "AUTHENTICATION",
-    "nodes": [
-        {
-            "id": "start",
-            "type": "START",
-            "onSuccess": "sso_check"
-        },
-        {
-            "id": "sso_check",
-            "type": "TASK_EXECUTION",
-            "executor": {
-                "name": "SSOCheckExecutor"
-            },
-            "properties": {
-                "checkpointRef": "session"
-            },
-            "onSuccess": "session",
-            "onFailure": "prompt_credentials"
-        },
-        {
-            "id": "basic_auth",
-            "type": "TASK_EXECUTION",
-            "executor": {
-                "name": "CredentialsAuthExecutor"
-            },
-            "onSuccess": "session",
-            "onIncomplete": "prompt_credentials"
-        },
-        {
-            "id": "prompt_credentials",
-            "type": "PROMPT",
-            "meta": {
-                "components": [
-                    {
-                        "align": "center",
-                        "type": "TEXT",
-                        "id": "text_001",
-                        "label": "{{ t(signin:forms.credentials.title) }}",
-                        "variant": "HEADING_1"
-                    },
-                    {
-                        "type": "BLOCK",
-                        "id": "block_001",
-                        "components": [
-                            {
-                                "id": "input_001",
-                                "ref": "username",
-                                "type": "TEXT_INPUT",
-                                "label": "{{ t(signin:forms.credentials.fields.username.label) }}",
-                                "required": true,
-                                "placeholder": "{{ t(signin:forms.credentials.fields.username.placeholder) }}"
-                            },
-                            {
-                                "id": "input_002",
-                                "ref": "password",
-                                "type": "PASSWORD_INPUT",
-                                "label": "{{ t(signin:forms.credentials.fields.password.label) }}",
-                                "required": true,
-                                "placeholder": "{{ t(signin:forms.credentials.fields.password.placeholder) }}"
-                            },
-                            {
-                                "type": "ACTION",
-                                "id": "action_001",
-                                "label": "{{ t(signin:forms.credentials.actions.submit.label) }}",
-                                "variant": "PRIMARY",
-                                "eventType": "SUBMIT"
-                            }
-                        ]
-                    }
-                ]
-            },
-            "prompts": [
-                {
-                    "inputs": [
-                        {
-                            "ref": "input_001",
-                            "identifier": "username",
-                            "type": "TEXT_INPUT",
-                            "required": true
-                        },
-                        {
-                            "ref": "input_002",
-                            "identifier": "password",
-                            "type": "PASSWORD_INPUT",
-                            "required": true
-                        }
-                    ],
-                    "action": {
-                        "ref": "action_001",
-                        "nextNode": "basic_auth"
-                    }
-                }
-            ]
-        },
-        {
-            "id": "session",
-            "type": "TASK_EXECUTION",
-            "executor": {
-                "name": "SessionExecutor"
-            },
-            "onSuccess": "authorization_check"
-        },
-        {
-            "id": "authorization_check",
-            "type": "TASK_EXECUTION",
-            "executor": {
-                "name": "AuthorizationExecutor"
-            },
-            "onSuccess": "auth_assert"
-        },
-        {
-            "id": "auth_assert",
-            "type": "TASK_EXECUTION",
-            "executor": {
-                "name": "AuthAssertExecutor"
-            },
-            "onSuccess": "end"
-        },
-        {
-            "id": "end",
-            "type": "END"
-        }
-    ]
-}
